Added '-pause' option to BaseCommandLineTool to force a pause after setup() is complete, before continuing with run(). This is targeted specifically at connecting a profiling tool after completing setup() to collect statistics on the main execution path.
When printing usage output, if the user specified -? or -help along
with other options, re-initialize an instance of BaseCommandlineTool
so that the default=... output matches the real default (and not the
arguments specified along with -?).

Moved configuration of GlobalConfigProperties to before command-line
parsing, so that GlobalConfigProperties is available during class
initalization (e.g., when creating enum instances).

Allowed skipping one or more header lines in inputLines() and other
convenience methods.